+++ src/table/base/TableRow.tsx
@@ -5,7 +5,7 @@ import { DragSortInnerProps } from '../../_util/useDragSorter';
 import { useTableContext } from './TableContext';
 import TableCell from './TableCell';
 import { DataType, TdBaseTableProps, RowspanColspan, RowspanAndColspanParams, TdPrimaryTableProps } from '../type';
-import { RowSkipTdSpanColIndexsMap } from './TableBody';
+import { RowSkipTdSpanColIndexsMap, RowEvents } from './TableBody';
 import { ExpandInnerProps } from './Table';
 
 interface MergeCellsProps {
@@ -25,35 +25,8 @@ interface RowProps<D extends DataType> extends MergeCellsProps, ExpandProps, Dra
   rowClassName?: TdBaseTableProps['rowClassName'];
   rowIndex?: number;
   rowKey: TdBaseTableProps['rowKey'];
+  rowEvents: RowEvents;
 }
-
-type RowEventName =
-  | 'onClick'
-  | 'onDoubleClick'
-  | 'onMouseOver'
-  | 'onMouseDown'
-  | 'onMouseEnter'
-  | 'onMouseLeave'
-  | 'onMouseUp';
-type APIRowEventName =
-  | 'onRowClick'
-  | 'onRowDbClick'
-  | 'onRowHover'
-  | 'onRowMousedown'
-  | 'onRowMouseenter'
-  | 'onRowMouseleave'
-  | 'onRowMouseup';
-export type RowEvents = Record<RowEventName, React.MouseEventHandler<HTMLElement>> | { [key: string]: Function };
-
-const rowEventsMap: Record<RowEventName, APIRowEventName> = {
-  onClick: 'onRowClick',
-  onDoubleClick: 'onRowDbClick',
-  onMouseOver: 'onRowHover',
-  onMouseDown: 'onRowMousedown',
-  onMouseEnter: 'onRowMouseenter',
-  onMouseLeave: 'onRowMouseleave',
-  onMouseUp: 'onRowMouseup',
-};
 
 const TableRow = <D extends DataType>(props: RowProps<D>) => {
   const {
@@ -68,6 +41,7 @@ const TableRow = <D extends DataType>(props: RowProps<D>) => {
     expandOnRowClick,
     handleExpandChange,
     sortOnRowDraggable,
+    rowEvents,
     onDragStart,
     onDragOver,
     onDrop,
@@ -115,8 +89,6 @@ const TableRow = <D extends DataType>(props: RowProps<D>) => {
   if (typeof rowClassName === 'function') {
     classes = rowClassName({ row: record, rowIndex });
   }
-
-  const rowEvents = getRowEvents(record, rowIndex);
 
   function getCustomRender({ record, colKey, cell, render }) {
     if (typeof cell === 'string' || isValidElement(cell)) {
@@ -275,24 +247,6 @@ const TableRow = <D extends DataType>(props: RowProps<D>) => {
       };
     }
     return {};
-  }
-
-  function getRowEvents(row, index): RowEvents {
-    const rowEventProps = {};
-    Object.keys(rowEventsMap).forEach((eventName) => {
-      const apiEventName = rowEventsMap[eventName];
-      const apiEvent = props[apiEventName];
-      if (apiEvent) {
-        rowEventProps[eventName] = (e) => {
-          apiEvent({
-            row,
-            index,
-            e,
-          });
-        };
-      }
-    });
-    return rowEventProps;
   }
 
   return (
